requirements.txt
Django==3.0.10
django-heroku==0.3.1
django-cors-headers==3.4.0
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2-binary==2.8.6